// src/types/data.ts
import type { SelectOptionGroup, Vechile } from './filter';

export type DailyViews = {
	views: number;
	product_id: number;
	insert_Date: string;
};

export type ProductItem = {
	car_id: number;
	rent_car_id: number;
	status_id: number;
	user_id: number;
	dealer_user_id: number;
	paid_add: number;
	photo: string;
	pic_number: number;
	prod_year: number;
	prod_month: number;
	man_id: number;
	car_model: string;
	price: number;
	price_usd: number;
	first_deposit: number;
	price_value: number;
	fuel_type_id: number;
	gear_type_id: number;
	drive_type_id: number;
	door_type_id: number;
	color_id: number;
	saloon_color_id: number;
	cylinders: number;
	car_run: number;
	car_run_km: number;
	car_run_dim: number;
	engine_volume: number;
	airbags: number;
	abs: boolean;
	esd: boolean;
	el_windows: boolean;
	conditioner: boolean;
	leather: boolean;
	disks: boolean;
	nav_system: boolean;
	central_lock: boolean;
	hatch: boolean;
	right_wheel: boolean;
	alarm: boolean;
	board_comp: boolean;
	hydraulics: boolean;
	chair_warming: boolean;
	climat_control: boolean;
	obstacle_indicator: boolean;
	customs_passed: boolean;
	client_name: string;
	client_phone: string;
	model_id: number;
	location_id: number;
	parent_loc_id: number;
	tech_inspection: boolean;
	checked_for_duplicates: boolean;
	order_number: number;
	stickers: null | string[];
	changable: boolean;
	auction: boolean;
	has_turbo: boolean;
	for_rent: boolean;
	rent_daily: boolean;
	rent_purchase: boolean;
	rent_insured: boolean;
	rent_driver: boolean;
	currency_id: number;
	vehicle_type: number;
	category_id: number;
	vin: string;
	user_type: null | number | string;
	prom_color: number;
	special_persons: boolean;
	back_camera: boolean;
	car_desc: string;
	order_date: string;
	video_url: string;
	hp: number;
	hours_used: number;
	photo_ver: number;
	checked: boolean;
	lang_type_id: number;
	el_starter: number;
	start_stop: boolean;
	trunk: boolean;
	windshield: boolean;
	inspected_in_greenway: boolean;
	license_number: string;
	words_checked: number;
	is_payd: boolean;
	condition_type_id: number;
	primary_damage_type: number;
	secondary_damage_type: number;
	auction_has_key: number;
	is_auction: number;
	saloon_material_id: number;
	map_lat: number;
	map_long: number;
	zoom: number;
	predicted_price: string;
	hdd: number;
	map_title: string;
	has_catalyst: number;
	driving_range: number;
	element_volume: number;
	trim_id: null | number;
	is_from_rent: null | boolean;
	tmp: string;
	views: number;
	dealerId: null | number;
	has_logo: null | boolean;
	logo_ver: null | number;
	active_ads: null | number;
	dealer_title: string;
	has360: number;
	has_predicted_price: boolean;
	pred_first_breakpoint: number;
	pred_second_breakpoint: number;
	pred_min_price: number;
	pred_max_price: number;
	comfort_features: number[];
	daily_views: DailyViews;
};

export type PaginationMeta = {
	total: number;
	per_page: number;
	current_page: number;
	last_page: number;
};

export type Manufacturer = {
	man_id: string;
	man_name: string;
	is_car: '1' | '0';
	is_spec: '1' | '0';
	is_moto: '1' | '0';
};

export type Category = {
	category_id: number;
	category_type: number;
	has_icon: 1 | 0;
	title: string;
	seo_title: string;
	vehicle_types: number[];
};

export type ModelItem = {
	model_id: number;
	man_id: number;
	model_name: string;
	model_group: string;
	sort_order: number;
	cat_man_id: number;
	cat_model_id: number;
	cat_modif_id: number;
	is_car: false;
	is_moto: false;
	is_spec: true;
	show_in_salons: 1 | 0;
	shown_in_slider: 1 | 0;
	is_synced: 1 | 0;
};

export type ManufacturersGroupedByVechile = Record<Vechile, Manufacturer[]>;

export type CategoriesGroupedByVehicleType = Record<Vechile, Category[]>;

export type ModelsGroupedByManufacturerAndVechile = Record<
	Vechile,
	SelectOptionGroup[]
>;

export type ProductsResponse = {
	data: {
		items: ProductItem[];
		meta: PaginationMeta;
	};
};